Fix some issue

+++ admin/delete-category.php
@@ -1,18 +1,22 @@
 <?php
 require 'config/database.php';
 
-if(isset($_GET['id'])) {
+if (isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
     //FOR LATER
     //update category_id of posts that belong to this category to id of uncategorized category
+    $update_query = "UPDATE posts SET category_id = NULL WHERE category_id=$id";
 
-
-
-    //delete category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
+    if (!mysqli_errno($connection)) {
+        //delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        if (!mysqli_errno($connection)) {
+            $update_result = mysqli_query($connection, $update_query);
+        }
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
 }
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
 die();
